requirements.txt
python-dotenv>=1.0.0
pyotp>=2.9.0
smartapi-python>=1.3.0
logzero>=1.7.0
requests>=2.31.0
websocket-client>=1.5.0
flask>=3.0.0
flask-cors>=4.0.0
